Rename to canvas abd cleanup
